+++ Controllers/UsersController.cs
@@ -1,8 +1,14 @@
-﻿using System.Data.Entity;
+﻿using System;
+using System.Collections.Generic;
+using System.Data;
+using System.Data.Entity;
 using System.Linq;
 using System.Net;
+using System.Web;
 using System.Web.Mvc;
+using ProcurementSystem;
 using ProcurementSystem.Models;
+using ProcurementSystem.Models.Enums; 
 
 namespace ProcurementSystem.Controllers
 {
@@ -33,17 +39,20 @@ namespace ProcurementSystem.Controllers
         }
 
         // GET: Users/Create
-
-        [Authorize(Roles = "АДМІНІСТРАТОР")]
         public ActionResult Create()
         {
+            ViewBag.Role = new SelectList(
+                Enum.GetValues(typeof(UserRole))
+                    .Cast<UserRole>()
+                    .Select(s => new { Id = (int)s, Name = s.ToString() }),
+                "Id",
+                "Name");
             return View();
         }
 
         // POST: Users/Create
         [HttpPost]
         [ValidateAntiForgeryToken]
-        [Authorize(Roles = "АДМІНІСТРАТОР")]
         public ActionResult Create([Bind(Include = "Id,Login,Password,Role")] User user)
         {
             if (ModelState.IsValid)
@@ -53,11 +62,18 @@ namespace ProcurementSystem.Controllers
                 return RedirectToAction("Index");
             }
 
+            ViewBag.Role = new SelectList(
+                Enum.GetValues(typeof(UserRole))
+                    .Cast<UserRole>()
+                    .Select(s => new { Id = (int)s, Name = s.ToString() }),
+                "Id",
+                "Name",
+                (int)user.Role); 
+
             return View(user);
         }
 
         // GET: Users/Edit/5
-        [Authorize(Roles = "АДМІНІСТРАТОР")]
         public ActionResult Edit(int? id)
         {
             if (id == null)
@@ -69,13 +85,21 @@ namespace ProcurementSystem.Controllers
             {
                 return HttpNotFound();
             }
+
+            ViewBag.Role = new SelectList(
+                Enum.GetValues(typeof(UserRole))
+                    .Cast<UserRole>()
+                    .Select(s => new { Id = (int)s, Name = s.ToString() }),
+                "Id",
+                "Name",
+                (int)user.Role); 
+
             return View(user);
         }
 
         // POST: Users/Edit/5
         [HttpPost]
         [ValidateAntiForgeryToken]
-        [Authorize(Roles = "АДМІНІСТРАТОР")]
         public ActionResult Edit([Bind(Include = "Id,Login,Password,Role")] User user)
         {
             if (ModelState.IsValid)
@@ -84,11 +108,19 @@ namespace ProcurementSystem.Controllers
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
+
+            ViewBag.Role = new SelectList(
+                Enum.GetValues(typeof(UserRole))
+                    .Cast<UserRole>()
+                    .Select(s => new { Id = (int)s, Name = s.ToString() }),
+                "Id",
+                "Name",
+                (int)user.Role);
+
             return View(user);
         }
 
         // GET: Users/Delete/5
-        [Authorize(Roles = "АДМІНІСТРАТОР")]
         public ActionResult Delete(int? id)
         {
             if (id == null)
@@ -106,10 +138,18 @@ namespace ProcurementSystem.Controllers
         // POST: Users/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        [Authorize(Roles = "АДМІНІСТРАТОР")]
         public ActionResult DeleteConfirmed(int id)
         {
+            bool hasOrders = db.Orders.Any(o => o.UserId == id);
+
             User user = db.Users.Find(id);
+
+            if (hasOrders)
+            {
+                ModelState.AddModelError("", "Неможливо видалити користувача, оскільки за ним закріплені заявки.");
+                return View(user); 
+            }
+
             db.Users.Remove(user);
             db.SaveChanges();
             return RedirectToAction("Index");
